add error handling

+++ app/lib/actions.ts
@@ -22,19 +22,29 @@ export async function createInvoice(formData: FormData) {
     amount: formData.get("amount"),
     status: formData.get("status"),
   });
+
   const amountInCents = amount * 100;
   const date = new Date().toISOString().split("T")[0];
 
-  await sql`
-    INSERT INTO invoices (customer_id, amount, status, date)
-    VALUES (${customerId}, ${amountInCents}, ${status}, ${date})
-  `;
+  try {
+    await sql`
+		 INSERT INTO invoices (customer_id, amount, status, date)
+		 VALUES (${customerId}, ${amountInCents}, ${status}, ${date})
+	  `;
+  } catch (error) {
+    console.error("Failed to create invoice:", error);
+    throw new Error("Database Error: Failed to Create Invoice.");
+  }
 
   revalidatePath("/dashboard/invoices");
   redirect("/dashboard/invoices");
 }
 
 export async function updateInvoice(id: string, formData: FormData) {
+  if (!id) {
+    throw new Error("Invalid Invoice ID.");
+  }
+
   const { customerId, amount, status } = UpdateInvoice.parse({
     customerId: formData.get("customerId"),
     amount: formData.get("amount"),
@@ -43,17 +53,31 @@ export async function updateInvoice(id: string, formData: FormData) {
 
   const amountInCents = amount * 100;
 
-  await sql`
-	  UPDATE invoices
-	  SET customer_id = ${customerId}, amount = ${amountInCents}, status = ${status}
-	  WHERE id = ${id}
-	`;
+  try {
+    await sql`
+		 UPDATE invoices
+		 SET customer_id = ${customerId}, amount = ${amountInCents}, status = ${status}
+		 WHERE id = ${id}
+	  `;
+  } catch (error) {
+    console.error("Failed to update invoice:", error);
+    throw new Error("Database Error: Failed to Update Invoice.");
+  }
 
   revalidatePath("/dashboard/invoices");
   redirect("/dashboard/invoices");
 }
 
 export async function deleteInvoice(id: string) {
-  await sql`DELETE FROM invoices WHERE id = ${id}`;
-  revalidatePath("/dashboard/invoices");
+  if (!id) {
+    throw new Error("Invalid Invoice ID.");
+  }
+
+  try {
+    await sql`DELETE FROM invoices WHERE id = ${id}`;
+    revalidatePath("/dashboard/invoices");
+  } catch (error) {
+    console.error("Failed to delete invoice:", error);
+    throw new Error("Database Error: Failed to Delete Invoice.");
+  }
 }
